requirements.txt
# AI Engineer 60-Day Challenge - Python Dependencies

# Core ML/AI Libraries
openai
langchain
langchain-community
langchain-openai

# Azure AI Services
azure-openai
azure-search-documents
azure-identity
azure-keyvault-secrets

# Vector Databases
chromadb
faiss-cpu
pinecone-client

# Data Processing
pandas
numpy
scikit-learn

# Web Framework & API
fastapi
uvicorn
pydantic
httpx

# Observability & Monitoring
prometheus-client
opentelemetry-api
opentelemetry-sdk

# Testing
pytest
pytest-asyncio
pytest-cov

# Utilities
python-dotenv
pyyaml
tenacity

sentence-transformers
faiss-gpu
rank_bm25

